\name{BayesSAE-package}
\alias{BayesSAE-package}
% \alias{BayesSAE}
\docType{package}
\title{Bayesian Data Analysis of Small Area Models
}
\description{The package provides a variety of functions to analysis several specific small area area-level 
models in Bayesian context. Function \code{BayesSAE} specifies the model and obtain MCMC posterior draws. 
\code{summary}, extract information from the returned object of class \code{BayesSAE}. \code{replication}
gives the replicated data which could be used for both numerical and graphical posterior checks. Function 
\code{mcmc} generates the \code{mcmc} object of class which could be used for further MCMC diagnosis.    
}
\details{
\tabular{ll}{
Package: \tab BayesSAE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-17\cr
License: \tab GPL-2\cr
}
This package provides functions for Bayesian analysis of small area models. 
}
\author{
Chengchun Shi

Maintainer: Chengchun Shi <3100102177@zju.edu.cn>
}
\references{
Bell, W. and Basel, W. and Cruse, C. and Dalzell, L. and Maples, J and O'Hara, B and Powers, D. (2007) 
\emph{ Use of ACS Data to Produce SAIPE Model-Based Estimates of Poverty for Counties,} U.S. Census official 
paper

Gelman, A. and Carlin, J. B. and Stern, H. S. and Rubin, D. B. (2006). \emph{ Bayesian Data Analysis,} 
CRC Press Company.

Hawalay, S. and Lahiriz, P. (2012). \emph{ Hierarchical Bayes Estimation of Poverty Rates,} U.S. Census 
companion paper

Rao, J. N. K. (2003) \emph{ Small Area Estimation}. John Wiley and Sons.

You, Y. and Chapman, B. (2006) Small Area Estimation Using Area Level Models and Estimated Sampling 
Variances. \emph{ Survey Methodology,} \bold{ 32:} 97-103.
}

